Flask: add GET /api/summary/participants
